requirements.txt
fastapi
uvicorn
httpx
pandas
chromadb[persist]
sentence-transformers
python-dotenv
